requirements.txt
numpy
scipy
pandas
PyYAML
Pillow
